Set PKG_CONFIG_PATH for cpan builds
For some reason a number of the aliens are not
picking up the system installs withou it.

We could do it for each build that needs it
but this is simpler overall as often the
aliens are deep in dependency trees.

This commit also corrects a config var name
in the config update code.