Update portableshell.bat to return exit code from Perl invocations.
This change modifies the non-interactive path in portableshell.bat
(where it invokes Perl with all provided arguments) to return the exit
code from the Perl process via "exit /b %ERRORLEVEL%" to the caller.

Prior to this change, using "portableshell.bat" in place of "perl.exe"
to run commands like:

  portableshell.bat script.pl arg1 arg2

would return with an exit code of zero, which is problematic if you
want to use the batch file in a makefile recipe or in some other
environment where you need to take additional actions based on the
Perl script's exit code.